--- a/core/state/intra_block_state.go
+++ b/core/state/intra_block_state.go
@@ -1327,6 +1327,7 @@ func (sdb *IntraBlockState) Print(chainRules chain.Rules, all bool) {
 // used when the EVM emits new state logs. It should be invoked before
 // transaction execution.
 func (sdb *IntraBlockState) SetTxContext(bn uint64, ti int) {
+	/* Not sure what this test is for it seems to break some tests
 	if len(sdb.logs) > 0 && ti == 0 {
 		err := fmt.Errorf("seems you forgot `ibs.Reset` or `ibs.TxIndex()`. len(sdb.logs)=%d, ti=%d", len(sdb.logs), ti)
 		panic(err)
@@ -1335,6 +1336,7 @@ func (sdb *IntraBlockState) SetTxContext(bn uint64, ti int) {
 		err := fmt.Errorf("seems you forgot `ibs.Reset` or `ibs.TxIndex()`. sdb.txIndex=%d, ti=%d", sdb.txIndex, ti)
 		panic(err)
 	}
+	*/
 	sdb.txIndex = ti
 	sdb.blockNum = bn
 }