--- a/core/state/journal.go
+++ b/core/state/journal.go
@@ -20,6 +20,8 @@
 package state
 
 import (
+	"fmt"
+
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/holiman/uint256"
 )
@@ -50,7 +52,6 @@ func newJournal() *journal {
 }
 func (j *journal) Reset() {
 	j.entries = j.entries[:0]
-	//j.dirties = make(map[common.Address]int, len(j.dirties)/2)
 	clear(j.dirties)
 }
 
@@ -101,39 +102,40 @@ type (
 		prev    *stateObject
 	}
 	selfdestructChange struct {
-		account     common.Address
+		account     *common.Address
 		prev        bool // whether account had already selfdestructed
 		prevbalance uint256.Int
 	}
 
 	// Changes to individual accounts.
 	balanceChange struct {
-		account common.Address
+		account *common.Address
 		prev    uint256.Int
 	}
 	balanceIncrease struct {
-		account  common.Address
+		account  *common.Address
 		increase uint256.Int
 	}
 	balanceIncreaseTransfer struct {
 		bi *BalanceIncrease
 	}
 	nonceChange struct {
-		account common.Address
+		account *common.Address
 		prev    uint64
 	}
 	storageChange struct {
-		account  common.Address
-		key      common.Hash
-		prevalue uint256.Int
+		account     *common.Address
+		key         common.Hash
+		prevalue    uint256.Int
+		wasCommited bool
 	}
 	fakeStorageChange struct {
-		account  common.Address
+		account  *common.Address
 		key      common.Hash
 		prevalue uint256.Int
 	}
 	codeChange struct {
-		account  common.Address
+		account  *common.Address
 		prevcode []byte
 		prevhash common.Hash
 	}
@@ -191,7 +193,7 @@ func (ch resetObjectChange) dirtied() *common.Address {
 }
 
 func (ch selfdestructChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
 	}
@@ -199,11 +201,22 @@ func (ch selfdestructChange) revert(s *IntraBlockState) error {
 		obj.selfdestructed = ch.prev
 		obj.setBalance(ch.prevbalance)
 	}
+	if s.versionMap != nil {
+		if obj.original.Balance == ch.prevbalance {
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: BalancePath})
+		} else {
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: BalancePath}]; ok {
+				v.Val = ch.prev
+			}
+		}
+		s.versionedWrites.Delete(*ch.account, AccountKey{Path: SelfDestructPath})
+	}
+
 	return nil
 }
 
 func (ch selfdestructChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 var ripemd = common.HexToAddress("0000000000000000000000000000000000000003")
@@ -215,31 +228,44 @@ func (ch touchChange) revert(s *IntraBlockState) error {
 func (ch touchChange) dirtied() *common.Address { return &ch.account }
 
 func (ch balanceChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
 	}
+	if traceAccount(*ch.account) {
+		fmt.Printf("Revert Balance %x: %d, prev: %d, orig: %d\n", *ch.account, obj.data.Balance, ch.prev, obj.original.Balance)
+	}
 	obj.setBalance(ch.prev)
+	if s.versionMap != nil {
+		if obj.original.Balance == ch.prev {
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: BalancePath})
+		} else {
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: BalancePath}]; ok {
+				v.Val = ch.prev
+			}
+		}
+	}
+
 	return nil
 }
 
 func (ch balanceChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch balanceIncrease) revert(s *IntraBlockState) error {
-	if bi, ok := s.balanceInc[ch.account]; ok {
+	if bi, ok := s.balanceInc[*ch.account]; ok {
 		bi.increase.Sub(&bi.increase, &ch.increase)
 		bi.count--
 		if bi.count == 0 {
-			delete(s.balanceInc, ch.account)
+			delete(s.balanceInc, *ch.account)
 		}
 	}
 	return nil
 }
 
 func (ch balanceIncrease) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch balanceIncreaseTransfer) dirtied() *common.Address {
@@ -251,46 +277,79 @@ func (ch balanceIncreaseTransfer) revert(s *IntraBlockState) error {
 	return nil
 }
 func (ch nonceChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
 	}
 	obj.setNonce(ch.prev)
+	if s.versionMap != nil {
+		if obj.original.Nonce == ch.prev {
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: NoncePath})
+		} else {
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: NoncePath}]; ok {
+				v.Val = ch.prev
+			}
+		}
+	}
+
 	return nil
 }
 
 func (ch nonceChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch codeChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
 	}
 	obj.setCode(ch.prevhash, ch.prevcode)
+	if s.versionMap != nil {
+		if obj.original.CodeHash == ch.prevhash {
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: CodePath})
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: CodeHashPath})
+		} else {
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: CodePath}]; ok {
+				v.Val = ch.prevcode
+			}
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: CodeHashPath}]; ok {
+				v.Val = ch.prevhash
+			}
+		}
+	}
 	return nil
 }
 
 func (ch codeChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch storageChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
+	}
+
+	if s.versionMap != nil {
+		if ch.wasCommited {
+			s.versionedWrites.Delete(*ch.account, AccountKey{Path: StatePath, Key: ch.key})
+		} else {
+			if v, ok := s.versionedWrites[*ch.account][AccountKey{Path: StatePath, Key: ch.key}]; ok {
+				v.Val = ch.prevalue
+			}
+		}
 	}
 	obj.setState(ch.key, ch.prevalue)
 	return nil
 }
 
 func (ch storageChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch fakeStorageChange) revert(s *IntraBlockState) error {
-	obj, err := s.getStateObject(ch.account)
+	obj, err := s.getStateObject(*ch.account)
 	if err != nil {
 		return err
 	}
@@ -299,7 +358,7 @@ func (ch fakeStorageChange) revert(s *IntraBlockState) error {
 }
 
 func (ch fakeStorageChange) dirtied() *common.Address {
-	return &ch.account
+	return ch.account
 }
 
 func (ch transientStorageChange) revert(s *IntraBlockState) error {
@@ -321,6 +380,9 @@ func (ch refundChange) dirtied() *common.Address {
 }
 
 func (ch addLogChange) revert(s *IntraBlockState) error {
+	if ch.txIndex >= len(s.logs) {
+		panic(fmt.Sprintf("can't revert log index %v, max: %v", ch.txIndex, len(s.logs)-1))
+	}
 	txnLogs := s.logs[ch.txIndex]
 	s.logs[ch.txIndex] = txnLogs[:len(txnLogs)-1] // revert 1 log
 	if len(s.logs[ch.txIndex]) == 0 {