… cmd/devnet/services/polygon/proofgenerator_test.go
… core/state/access_list_test.go
… core/state/database_test.go
… core/state/intra_block_state_logger_test.go
… core/state/intra_block_state_test.go
… core/state/state_test.go
… core/state/versionedio.go
… core/state_transition.go
… core/tracing/hooks.go
… core/vm/evm.go
… core/vm/evmtypes/evmtypes.go
… core/vm/instructions.go
… core/vm/interpreter.go
… eth/stagedsync/stage_mining_exec.go
… eth/tracers/internal/tracetest/calltrace_test.go
… execution/abi/bind/backends/simulated.go
… execution/consensus/aura/aura.go
… execution/consensus/consensus.go
… execution/consensus/ethash/consensus.go
… execution/consensus/merge/merge.go
… execution/exec3/historical_trace_worker.go
… execution/exec3/state.go
… execution/exec3/trace_worker.go
… polygon/aa/aa_gas.go
… polygon/bor/bor.go
… rpc/jsonrpc/eth_callMany.go
… rpc/jsonrpc/otterscan_search_trace.go
… rpc/jsonrpc/overlay_api.go
… rpc/jsonrpc/receipts/receipts_generator.go
… rpc/jsonrpc/trace_adhoc.go