Versioned IBS is used by parallel processing to keep track of reads and writes between parallel threads running on the same block.

If the versionedMap is set on the IBS it will use this to mediate read and write activity to make sure that there dependency clash between transactions running in parallel.  If a dependency is encountered execution will be stopped with a dependency panic which is handled by the parallel scheduler which re-schedules execution after the dependencies have been resolves.

Note that this version of the IBS has a block in in its context.  This is not functional it is purely used for tracing.   This is because differences in state setting can have subtle side effects which can cause issues several blocks later.  Having the block in the trace output makes it easier to tie together cause and effect.