@mh0lt 2025-06-02
Fixed a number of performance regressions:  Need additional checks on final merge:

```
mholt@mhgig1:~/erigon/erigon$ ~/go/bin/benchstat old-simple.txt new-simple.txt
goos: linux
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                                       │ old-simple.txt │           new-simple.txt           │
                                       │     sec/op     │   sec/op     vs base               │
SimpleLoop/staticcall-identity-100M-16      294.7m ± 3%   299.0m ± 2%       ~ (p=0.247 n=10)
SimpleLoop/call-identity-100M-16            362.0m ± 4%   370.3m ± 2%  +2.29% (p=0.019 n=10)
SimpleLoop/loop-100M-16                     270.3m ± 3%   267.0m ± 3%       ~ (p=0.247 n=10)
SimpleLoop/loop2-100M-16                    322.9m ± 6%   319.7m ± 4%       ~ (p=0.481 n=10)
SimpleLoop/call-nonexist-100M-16            305.8m ± 1%   316.9m ± 7%  +3.62% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-16                 334.1m ± 3%   341.1m ± 2%  +2.11% (p=0.015 n=10)
SimpleLoop/call-reverting-100M-16           527.2m ± 2%   523.0m ± 1%       ~ (p=0.353 n=10)
geomean                                     337.8m        340.8m       +0.90%

                                       │ old-simple.txt │            new-simple.txt             │
                                       │      B/op      │     B/op       vs base                │
SimpleLoop/staticcall-identity-100M-16    67.94Mi ±  0%   83.61Mi ±  0%  +23.07% (p=0.000 n=10)
SimpleLoop/call-identity-100M-16          66.57Mi ±  0%   81.93Mi ±  0%  +23.07% (p=0.000 n=10)
SimpleLoop/loop-100M-16                   2.107Ki ±  6%   1.578Ki ±  0%  -25.12% (p=0.000 n=10)
SimpleLoop/loop2-100M-16                  2.109Ki ± 20%   1.578Ki ± 15%  -25.19% (p=0.001 n=10)
SimpleLoop/call-nonexist-100M-16          22.78Mi ±  0%   22.78Mi ±  0%   -0.00% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-16               22.78Mi ±  0%   39.86Mi ±  0%  +74.99% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-16         201.7Mi ±  0%   218.0Mi ±  0%   +8.11% (p=0.000 n=10)
geomean                                   2.963Mi         3.170Mi         +7.00%

                                       │ old-simple.txt │            new-simple.txt             │
                                       │   allocs/op    │  allocs/op    vs base                 │
SimpleLoop/staticcall-identity-100M-16      2.740M ± 0%    3.425M ± 0%   +25.00% (p=0.000 n=10)
SimpleLoop/call-identity-100M-16            2.685M ± 0%    3.356M ± 0%   +25.00% (p=0.000 n=10)
SimpleLoop/loop-100M-16                     13.000 ± 0%     7.000 ± 0%   -46.15% (p=0.000 n=10)
SimpleLoop/loop2-100M-16                    13.000 ± 0%     7.000 ± 0%   -46.15% (p=0.000 n=10)
SimpleLoop/call-nonexist-100M-16            746.3k ± 0%    746.3k ± 0%    -0.00% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-16                 746.3k ± 0%   1492.5k ± 0%  +100.00% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-16           3.571M ± 0%    4.286M ± 0%   +20.00% (p=0.000 n=10)
geomean                                     58.94k         59.65k         +1.20%
```